refactor(task-engine): move lifecycle/cleanup/terminals + flip the barrel [SHR-195]
Move startTask/addAgent/resumeTask/hopAgent/preflightWorktree into
task-engine/lifecycle.ts, close/softDelete/delete/stopAgent into cleanup.ts,
terminal helpers into terminals.ts. task-engine/index.ts is now the facade;
server/task-runner.ts is a 6-line back-compat barrel (export * + two export
type). All caller import paths unchanged. Ordering invariants preserved
(setTmuxSession after session create; unlink worktree before delete).

startTask kept as a cohesive stage-marked body (not split into 8 sub-fns) to
avoid risky parameter-threading in the highest-risk path.

Modular 04 step 04.4.
